requirements.txt
-i https://pypi.org/simple
amqp==5.3.1; python_version >= '3.6'
anyio==4.9.0; python_version >= '3.9'
arrow==1.3.0; python_version >= '3.8'
asgiref==3.8.1; python_version >= '3.8'
attrs==25.3.0; python_version >= '3.8'
autobahn==24.4.2; python_version >= '3.9'
automat==24.8.1; python_version >= '3.8'
billiard==4.2.1; python_version >= '3.7'
blinker==1.9.0; python_version >= '3.9'
boto3==1.38.23; python_version >= '3.9'
botocore==1.38.23; python_version >= '3.9'
captcha==0.7.1; python_version >= '3.8'
cassandra-driver==3.29.2
celery==5.4.0; python_version >= '3.8'
certifi==2025.4.26; python_version >= '3.6'
cffi==1.17.1; python_version >= '3.8'
channels==4.2.2; python_version >= '3.8'
channels-redis==4.2.1; python_version >= '3.8'
charset-normalizer==3.4.1; python_version >= '3.7'
click==8.1.8; python_version >= '3.7'
click-didyoumean==0.3.1; python_full_version >= '3.6.2'
click-plugins==1.1.1
click-repl==0.3.0; python_version >= '3.6'
cloudinary==1.44.0
colorama==0.4.6; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3, 3.4, 3.5, 3.6'
constantly==23.10.4; python_version >= '3.8'
cron-descriptor==1.4.5
crontab==1.0.4
cryptography==44.0.2; python_version >= '3.7' and python_full_version not in '3.9.0, 3.9.1'
daphne==4.1.2; python_version >= '3.8'
defusedxml==0.7.1; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3, 3.4'
dj-database-url==2.3.0
django==5.2.3; python_version >= '3.10'
django-cassandra-engine==1.10.0; python_version >= '3.10' and python_version < '4.0'
django-celery-beat==2.8.1; python_version >= '3.8'
django-celery-results==2.6.0
django-cloudinary-storage==0.3.0
django-cors-headers==4.7.0; python_version >= '3.9'
django-countries==7.6.1
django-db-geventpool==4.0.8
django-debug-toolbar==5.2.0; python_version >= '3.9'
django-filter==25.1; python_version >= '3.9'
django-graphql-jwt==0.4.0; python_version >= '3.6'
django-ratelimit==4.1.0; python_version >= '3.7'
django-redis==5.4.0; python_version >= '3.6'
django-rq==3.0.1
django-rq-dashboard==0.3.3
django-storages==1.14.6; python_version >= '3.7'
django-templated-mail==1.1.1
django-timezone-field==7.1; python_version >= '3.8' and python_version < '4.0'
djangorestframework==3.16.0; python_version >= '3.9'
djangorestframework-simplejwt==5.5.0; python_version >= '3.9'
djoser==2.3.1; python_version >= '3.8' and python_version < '4.0'
dnspython==2.7.0; python_version >= '3.9'
drf-nested-routers==0.94.2; python_version >= '3.8'
drf-spectacular==0.28.0; python_version >= '3.7'
drf-yasg==1.21.10; python_version >= '3.6'
factory-boy==3.3.3; python_version >= '3.8'
faker==37.3.0; python_version >= '3.9'
flask==3.1.1; python_version >= '3.9'
flower==2.0.1; python_version >= '3.7'
freezegun==1.5.2; python_version >= '3.8'
geomet==0.2.1.post1; python_version >= '2.7' and python_version != '3.3' and python_version < '4'
graphene==3.4.3
graphene-django==3.2.3
graphql-core==3.2.6; python_version >= '3.6' and python_version < '4'
graphql-relay==3.2.0; python_version >= '3.6' and python_version < '4'
gunicorn==23.0.0; python_version >= '3.7'
h11==0.14.0; python_version >= '3.7'
h2==4.2.0; python_version >= '3.9'
hpack==4.1.0; python_version >= '3.9'
httpcore==1.0.8; python_version >= '3.8'
httpx==0.28.1; python_version >= '3.8'
humanize==4.12.3; python_version >= '3.9'
hyperframe==6.1.0; python_version >= '3.9'
hyperlink==21.0.0
idna==3.10; python_version >= '3.6'
incremental==24.7.2; python_version >= '3.8'
inflection==0.5.1; python_version >= '3.5'
iniconfig==2.1.0; python_version >= '3.8'
itsdangerous==2.2.0; python_version >= '3.8'
jinja2==3.1.6; python_version >= '3.7'
jmespath==1.0.1; python_version >= '3.7'
jsonschema==4.23.0; python_version >= '3.8'
jsonschema-specifications==2024.10.1; python_version >= '3.9'
kombu==5.5.0; python_version >= '3.8'
markupsafe==3.0.2; python_version >= '3.9'
msgpack==1.1.0; python_version >= '3.8'
nest-asyncio==1.6.0; python_version >= '3.5'
oauthlib==3.2.2; python_version >= '3.6'
packaging==25.0; python_version >= '3.8'
pillow==11.1.0; python_version >= '3.9'
pluggy==1.6.0; python_version >= '3.9'
priority==1.3.0
prometheus-client==0.22.0; python_version >= '3.9'
promise==2.3
prompt-toolkit==3.0.50; python_full_version >= '3.8.0'
psycogreen==1.0.2
psycopg==3.2.9; python_version >= '3.8'
psycopg2-binary==2.9.10; python_version >= '3.8'
pyasn1==0.6.1; python_version >= '3.8'
pyasn1-modules==0.4.1; python_version >= '3.8'
pycparser==2.22; python_version >= '3.8'
pyjwt==2.9.0; python_version >= '3.8'
pyopenssl==25.0.0; python_version >= '3.7'
pytest==8.3.5; python_version >= '3.8'
pytest-django==4.11.1; python_version >= '3.8'
pytest-factoryboy==2.7.0; python_version >= '3.8'
pytest-mock==3.14.0; python_version >= '3.8'
python-crontab==3.2.0
python-dateutil==2.9.0.post0; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3'
python-dotenv==1.1.0; python_version >= '3.9'
python-telegram-bot==22.2; python_version >= '3.9'
python3-openid==3.2.0
pytz==2025.2
pyyaml==6.0.2; python_version >= '3.8'
qrcode==8.2; python_version >= '3.9' and python_version < '4.0'
redis==6.1.0; python_version >= '3.8'
redis-sentinel-url==1.0.1
referencing==0.36.2; python_version >= '3.9'
requests==2.32.3; python_version >= '3.8'
requests-oauthlib==2.0.0; python_version >= '3.4'
rpds-py==0.23.1; python_version >= '3.9'
rq==2.4.0; python_version >= '3.9'
rq-dashboard==0.8.3.2
rq-scheduler==0.14.0
s3transfer==0.13.0; python_version >= '3.9'
scylla-driver==3.29.0
service-identity==24.2.0; python_version >= '3.8'
setuptools==80.9.0; python_version >= '3.9'
six==1.17.0; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3'
sniffio==1.3.1; python_version >= '3.7'
social-auth-app-django==5.4.3; python_version >= '3.9'
social-auth-core==4.5.6; python_version >= '3.9'
sqlparse==0.5.3; python_version >= '3.8'
strawberry-django-plus==3.1.1; python_version >= '3.8' and python_version < '4.0'
strawberry-graphql==0.278.0; python_version >= '3.9' and python_version < '4.0'
strawberry-graphql-django==0.9.5; python_version >= '3.7' and python_version < '4.0'
text-unidecode==1.3
tornado==6.5.1; python_version >= '3.9'
twisted[tls]==24.11.0; python_full_version >= '3.8.0'
txaio==23.1.1; python_version >= '3.7'
types-python-dateutil==2.9.0.20250516; python_version >= '3.9'
typing-extensions==4.12.2; python_version >= '3.8'
tzdata==2025.1; python_version >= '2'
uritemplate==4.1.1; python_version >= '3.6'
urllib3==2.3.0; python_version >= '3.9'
uvicorn==0.34.3; python_version >= '3.9'
vine==5.1.0; python_version >= '3.6'
wcwidth==0.2.13
websockets==15.0.1; python_version >= '3.9'
werkzeug==3.1.3; python_version >= '3.9'
whitenoise==6.9.0; python_version >= '3.9'
zope.interface==7.2; python_version >= '3.8'
